requirements.txt
agate==1.7.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
babel==2.18.0
certifi==2026.2.25
cffi==1.17.1
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
daff==1.4.2
dbt-adapters==1.22.6
dbt-common==1.37.2
dbt-core==1.7.19
dbt-duckdb==1.7.4
dbt-extractor==0.5.1
dbt-protos==1.0.431
dbt-semantic-interfaces==0.4.4
deepdiff==7.0.1
duckdb==1.4.4
fastapi==0.133.1
idna==3.11
importlib-metadata==6.11.0
isodate==0.6.1
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
leather==0.4.1
Logbook==1.5.3
MarkupSafe==3.0.3
mashumaro==3.14
minimal-snowplow-tracker==0.0.2
more-itertools==10.8.0
msgpack==1.1.2
networkx==3.6.1
ordered-set==4.1.0
orderly-set==5.5.0
packaging==26.0
parsedatetime==2.6
pathspec==0.11.2
protobuf==4.25.8
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
snowplow-tracker==1.1.0
sqlparse==0.5.4
starlette==0.52.1
text-unidecode==1.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.20
zipp==3.23.0

uvicorn[standard]==0.30.0
httpx==0.27.0
pytest==8.3.0